refactor: 경험 기간 포맷 유틸 분리
